+++ test/Data/Text/Extras.hs
@@ -1,0 +1,7 @@
+module Data.Text.Extras where
+
+import Data.Text
+import Test.QuickCheck
+
+instance Arbitrary Text where
+  arbitrary = pack <$> arbitrary